@@ -1,6 +1,6 @@
 module github.com/ooni/oohttp
 
-go 1.17
+go 1.18
 
 require golang.org/x/net v0.0.0-20220412020605-290c469a71a5
 
